Fixed name on connect new repo and moved it to module root
